The bot will now restart itself on tcp_error or tcp_closed
For this change I had to carry the starting Args in the loop so that I can
restart the bot if needed with the same arguments.

It also usues a quadratic backoff algorithm so that it will wait 5 sec on
the first restart, 20 on the second, 45 on the third, up to 125 on the fifth
and more.