Better nickname handling in the tell plugin
Before this the 'tell' plugin considered a nickname to be everything up to the
first space. That broke on !tell nickname: since it would consider the : part
of the nickname, even when it's forbiden.

With this change, the regexp now explicitly detects/splits on non-nickname chars.
(also added a comment for the unfinished fancy_time function)