Improvements to the Makefile
Now the Makfile will automatically find all *.erl files in ./src/ and compile
them to the same directory structure to ./ebin/

I use a shell invocation of 'find' to find the source files, because I didn't
know if Makefile itself could do it.

I did this change in order to be able to easily create new subdirectories for
more complicated plugins.