Patch a little bug when reconnecting
The "connecting" state can be entered from 2 previous states. The first is when in
offline state the FSM gets the explicit "connect" event, then connecting needs
to signal back the success of the connection. The second is when in online
state the connection is broken, then it doesn't signal anything.