Finally make debug to stdout robust
Erlang's io:* functions are all borked beyond comprehension.
Not only that they fail to write a binary with non-latin1 chars
they'll not throw a proper exception or anything.

That ment the bot would hard block when connecting to the
lindbohm.freenode.net irc server since the use latin1 in the
welcome message (Dénis Lindbohm).