avoid a concurrency issue when the connection dies
Immediately switch to standby state when the connection process dies. This avoids the situation
when the bot tries to send something on a now 'undefined' connection (Conn) which kills the fsm.

In standby state it never tries to send anything.

this is obviously an old bug, and this commit makes the code a bit simpler, and
correct.