the bot should be a permanent process
Thsi was probably a sloppy copy/paste from somewhere. The childspec for the bot
connection needs to be set to permanent, so that the supervisor will restart
it when it dies.

I haven't noticed this since the ircbot_fsm has never died until I noticed the
race condition fixed in the previous commit.

One more issue remains though. When restarted the plugins process blocks
somewhere.