migrate from gen_fsm to gen_statem
This is a big refactor to use gen_statem instead of gen_fsm. gen_fsm is
now kind-of-deprecated in Erlang, and gen_statem is the replacement.

The gen_statem api is a bit better, even though it's also more powerful,
and the documentation needs some work on.

Fixes compile-time warnings.

update the README to use rebar3 mostly